exposed Manager::convertToArray method to Converter
